+++ src/main/java/com/jcloisterzone/game/phase/CocScoringPhase.java
@@ -5,10 +5,13 @@ import java.util.function.Function;
 import com.jcloisterzone.board.Location;
 import com.jcloisterzone.board.Position;
 import com.jcloisterzone.board.pointer.FeaturePointer;
+import com.jcloisterzone.feature.Cloister;
+import com.jcloisterzone.feature.Completable;
 import com.jcloisterzone.feature.Farm;
 import com.jcloisterzone.feature.Feature;
 import com.jcloisterzone.figure.Barn;
 import com.jcloisterzone.game.RandomGenerator;
+import com.jcloisterzone.game.capability.AbbeyCapability;
 import com.jcloisterzone.game.capability.BarnCapability;
 import com.jcloisterzone.game.capability.CountCapability;
 import com.jcloisterzone.game.state.GameState;
@@ -16,8 +19,6 @@ import com.jcloisterzone.game.state.PlacedTile;
 
 import io.vavr.Predicates;
 import io.vavr.Tuple2;
-import io.vavr.collection.HashMap;
-import io.vavr.collection.Map;
 
 @RequiredCapability(CountCapability.class)
 public class CocScoringPhase extends AbstractCocScoringPhase {
@@ -27,18 +28,24 @@ public class CocScoringPhase extends AbstractCocScoringPhase {
     }
 
     @Override
-    protected Map<Location, Function<Feature, Boolean>> getScoredQuarters(GameState state) {
-        Map<Location, Function<Feature, Boolean>> quarters;
-        quarters = HashMap.of(
-            Location.QUARTER_BLACKSMITH, null,
-            Location.QUARTER_CASTLE, null,
-            Location.QUARTER_CATHEDRAL, null
-        );
+    protected Function<Feature, Boolean> getAllowedFeaturesFilter(GameState state) {
+        PlacedTile lastPlaced = state.getLastPlaced();
+        Position lastPlacedPos = lastPlaced.getPosition();
 
+        java.util.Set<Completable> justPlacedAbbeyAdjacent = new java.util.HashSet<>();
+        if (AbbeyCapability.isAbbey(lastPlaced.getTile())) {
+            for (Tuple2<Location, PlacedTile> t : state.getAdjacentTiles2(lastPlacedPos)) {
+                PlacedTile pt = t._2;
+                Feature feature = state.getFeaturePartOf(new FeaturePointer(pt.getPosition(), t._1.rev()));
+                if (feature instanceof Completable) {
+                    justPlacedAbbeyAdjacent.add((Completable) feature);
+                }
+            }
+        }
+
+        java.util.HashSet<Feature> barnInvolvedFarms = new java.util.HashSet<>();
 
         if (state.getCapabilities().contains(BarnCapability.class)) {
-            java.util.HashSet<Feature> barnInvolvedFarms = new java.util.HashSet<>();
-
             FeaturePointer placedBarnPtr = state.getCapabilityModel(BarnCapability.class);
             Farm placedBarnFarm = placedBarnPtr == null ? null : (Farm) state.getFeature(placedBarnPtr);
             if (placedBarnFarm != null) {
@@ -46,7 +53,6 @@ public class CocScoringPhase extends AbstractCocScoringPhase {
 
             }
 
-            PlacedTile lastPlaced = state.getLastPlaced();
             Position pos = lastPlaced.getPosition();
             state.getTileFeatures2(pos)
                 .map(Tuple2::_2)
@@ -59,14 +65,40 @@ public class CocScoringPhase extends AbstractCocScoringPhase {
                 )
                 .filter(farm -> !farm.getFollowers(state).isEmpty())  //must contains at least one follower
                 .forEach(farm -> barnInvolvedFarms.add(farm));
-
-            if (!barnInvolvedFarms.isEmpty()) {
-                quarters = quarters.put(Location.QUARTER_MARKET, f -> barnInvolvedFarms.contains(f));
-            }
-
         }
 
-        return quarters;
-    }
+        return f -> {
+            if (f instanceof Farm) {
+                return barnInvolvedFarms.contains(f);
+            }
+            if (f instanceof Completable) {
+                Completable completable = (Completable) f;
+                if (!completable.isCompleted(state)) {
+                    return false;
+                }
 
+                if (justPlacedAbbeyAdjacent.contains(f)) {
+                    return true;
+                }
+
+                //feature lays on last placed tile -> is finished this turn
+                if (f.getPlaces().find(p -> p.getPosition().equals(lastPlacedPos)).isDefined()) {
+                    return true;
+                }
+
+                if (f instanceof Cloister) { // Cloister, not CloisterLike
+                    Position cloisterPos = f.getPlaces().get().getPosition();
+                    if (!Position.ADJACENT_AND_DIAGONAL
+                        .map(t -> cloisterPos.add(t._2))
+                        .filter(p -> p.equals(lastPlacedPos))
+                        .isEmpty()) {
+                        return true;
+                    }
+                }
+
+                return false;
+            }
+            throw new UnsupportedOperationException();
+        };
+    }
 }