+++ changelog.md
@@ -4,6 +4,8 @@
 
 ## upcoming release
 
+* The Count, final scoring change:
+  - followers can be moved from all quarters (not only from the Market to forms)
 * fix: road with bridge can remain not scored. This can happen when bridge was placed result of tile placement
   on adjacent tile (and only when tile bridge is rotated once). Such bridge can also cause error during final scoring.
 * fix: gold was not awarded from cloister tile itself (but only from adjacent tiles)